changed player wins to be Map<Integer,Double> to accomadate the possibility of storing money values

+++ src/edu/brown/cs32/MFTG/gui/gameboard/Board.java
@@ -466,7 +466,7 @@ public class Board extends JPanel {
 		}
 	}
 	
-	public void setWinnerData(Map<Integer,Integer> data){
+	public void setWinnerData(Map<Integer,Double> data){
 		//TODO Chizu can you implement this?
 	}
